Move modular inverse to independent file
